Add renders tasks test for tasksList comp & strongly type ergs
