+++ src/components/TasksList.vue
@@ -19,7 +19,7 @@ const { completeTask } = tasksStore
     <div class="">
       <div class="taskList">
         <div class="flex flex-col gap-y-2">
-          <div v-for="task in listTasks" :key="task.id" class="flex items-center bg-dark-300 rounded-md px-4 py-2.5">
+          <div v-for="task in listTasks" :key="task.id" class="flex items-center bg-dark-300 rounded-md px-4 py-2.5 task">
             <div class="cursor-pointer mr-2" @click="completeTask(task.id)">
               <Icon v-show="!task.completed" icon="ph-circle" />
               <Icon v-show="task.completed" icon="ph-check-circle" />
