+++ src/components/__tests__/TaskForm.spec.ts
@@ -2,6 +2,7 @@ import { createPinia, setActivePinia } from "pinia"
 import { beforeAll, describe, it, expect } from "vitest"
 import { mount } from "@vue/test-utils"
 import { useTaskStore } from "@/stores/task"
+import type { Task } from "@/stores/task"
 import TaskForm from "@/components/TaskForm.vue"
 
 describe("TaskForm", () => {
@@ -10,8 +11,7 @@ describe("TaskForm", () => {
   beforeAll(() => {
     setActivePinia(createPinia())
     store = useTaskStore()
-    store.selectedList = 1
-    store.nextId = 5
+    Object.assign(store, { selectedList: 1, nextId: 5 })
   })
 
   it("Renders properly", () => {
@@ -29,7 +29,7 @@ describe("TaskForm", () => {
     await wrapper.find("input").setValue("testTask")
     await wrapper.find("form").trigger("submit")
 
-    const task = store.tasks.find((task: any) => task.id === 5)
+    const task = store.tasks.find((task: Task) => task.id === 5)
     expect(task).toEqual({
       id: 5,
       title: "TestTask",
