--- src/components/__tests__/ToDoList.spec.ts
@@ -1,15 +1,0 @@
-import { createPinia, setActivePinia } from "pinia"
-import { beforeAll, describe, it, expect } from "vitest"
-import { mount } from "@vue/test-utils"
-import TasksList from "@/components/TasksList.vue"
-
-beforeAll(() => {
-  setActivePinia(createPinia())
-})
-
-describe("TasksList", () => {
-  it("renders properly", () => {
-    const wrapper = mount(TasksList)
-    expect(wrapper).toBeTruthy()
-  })
-})
